pin-1385-tests: Improve IT code
